move lightning stuff to a flask blueprint
