+++ templates/index.html
@@ -55,7 +55,7 @@
                     <p><center>User Order</center></p>
                 </div>
             </a>
-            <a href="{{ url_for('ln_ep') }}">
+            <a href="{{ url_for('ln_wallet.index') }}">
                 <div class='dashboard'>
                     <p><center>Lightning</center></p>
                 </div>
